Some matching cleanup and move a utility function out
